Misleading hasMany() relationship.
hasMany() relationship in Laravel refers to 1-n and doesn't require pivot table. While updating and creating hasMany() relationship in your model, this should be handled manually as explained in 'https://github.com/Laravel-Backpack/CRUD/issues/3031#issuecomment-660034201'. Referring this method through the documentation is misleading and better to be removed in my opinion.

There is also a relationship mentioned belongsToOne() which doesn't exist in Laravel core, removed this as well.