Update file name to: menu_items.blade.php
This line is described in detail at the bottom of this page and the file name there is the correct one that I can see in my fresh v6.0.3 backpack installation. Also, it appears that the menu items are created as components and can probably be used in many places, so reference to "sidebar menu" is probably no longer needed.