Rework div widget examples and explanation
I thought of adding the custom attribute to the first example, but then the first paragraph would have to be changed. Since that paragraph mentions including multiple widgets within the "div", I think it may be best to leave it as it is because it is a clear example for that.

I added the second example with the extra information and attributes to show how the widget can be used together with the generated output.