Fix typos and make minor edits
Fix typos and make minor edits

Both addon and add-ons was used in this file, and probably add-on with a hyphen is the better option to use. For example, see [add-on on Wikipedia](https://en.wikipedia.org/wiki/Add-on)

I added a semi-colon to a bullet point for consistency (as the other three bullet points had semi-colons already). However, it is probably acceptable not to use any punctuation at all in bullet points as it just makes writing documentation more simple.

The ... are not necessary and can be removed.