Fix typos and make minor edits
Fix typos and make minor edits.

The phrase "This B solution" seems to be in the wrong place or is no longer relevant, so I removed it. However, I am not sure about this.

I rewrote the sentences about why somebody might not like Backpack. I think it is better not to use that kind of negative phrasing in your documentation. The phrasing I have used is more neutral and concentrates of "use" rather than "not like".